+++ src/spring/azext_spring/application_accelerator.py
@@ -6,7 +6,7 @@
 from azure.cli.core.util import sdk_no_wait
 from knack.log import get_logger
 
-from .vendored_sdks.appplatform.v2022_11_01_preview import models
+from .vendored_sdks.appplatform.v2023_01_01_preview import models
 from .dev_tool_portal import (is_updatable as is_dev_tool_portal_updatable,
                               try_get as get_dev_tool_portal,
                               create_or_update as create_or_update_dev_tool_portal,